% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{est_sbetel}
\alias{est_sbetel}
\title{Estimate a smoothed BETEL model}
\usage{
est_sbetel(model, N, tune = "auto", itermax = 20, burn = NULL, verbose = TRUE)
}
\arguments{
\item{model}{A list returned by \code{init_sbetel()}, defining the model.}

\item{N}{Number of draws from the posterior distribution.}

\item{tune}{A parameter that controls the step size of the RWMH algorithm.
Defaults to \code{"auto"} in which case sufficiently efficient parameter value
is automatically chosen by mimicking the manual tuning process with \code{N = 100}.
Causion is advised when using this default setting.}

\item{itermax}{Maximum number of Newton-Rhapson iterations within the evaluation 
of the likelihood.}

\item{burn}{\code{NULL} or numerical value \code{> 0}. The length of the 
burn-in sample. Defaults to \code{NULL} in which case there is no burn-in.}

\item{verbose}{Logical. If \code{TRUE} messages are produced. Defaults to \code{TRUE}.}
}
\value{
\code{est_sbetel()} returns a list containing the output of the 
RWMH algorithm.
}
\description{
Estimates a smoothed BETEL model by generating a sample
from parameter posterior distribution by a RWMH algorithm.
}
\examples{
model_output <- est_sbetel(model, N = 1000)
}
