% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast_sbetel}
\alias{forecast_sbetel}
\title{Forecast with a smoothed (VAR) BETEL}
\usage{
forecast_sbetel(model, NN, horizon = 1, newdata = NULL, verbose = TRUE)
}
\arguments{
\item{model}{Estimated model generated by \code{est_sbetel}.}

\item{NN}{A number of draws from the predictive distribution.}

\item{horizon}{Forecasting horizon.}

\item{newdata}{A numerical matrix containing the predictive variables. 
Defaults to \code{NULL} in which case the end of the data used for 
estimation of the model is used.}

\item{verbose}{Logical. If \code{TRUE} messages are produced. Defaults to \code{TRUE}.}
}
\value{
\code{eval_sbetel()} returns an array of the form 
\code{(horizon, series, draw)} consisting of the draws from the
predictive distribution.
}
\description{
Samples from the predictive distribution of the given smoothed 
BETEL model. Supports only VAR-type models at the time.
}
\examples{
cast <- forecast_sbetel(model = model, NN = 1000, horizon = 8)
}
