% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{init_sbetel}
\alias{init_sbetel}
\title{Initialize a smoothed BETEL model}
\usage{
init_sbetel(
  y,
  g = "var",
  prior_fun = "var",
  bw = "auto",
  initial = NULL,
  args = list(),
  verbose = TRUE
)
}
\arguments{
\item{g}{A function of the form \code{g(th, y, args)} defining the model.
Inputs should include a numerical parameter vector \code{th}, a data matrix 
\code{y} and optional further arguments through a list called \code{args}.
Output shoud be a numerical matrix with its columns corresponding to the sample
moment conditions of the model. Defaults to \code{"var"} in which case a 
ready-made function for vector autoregressive models is used. For more, 
see details.}

\item{prior_fun}{A function that defines the prior data generating process. 
The only inputs should be the model generated by \code{init_sbetel()}. Further 
arguments to this function should be passed through \code{model$args}. 
Output should be a list with to elements called \code{yy} for dependent 
variables and \code{xx} for explanatory variables or a data matrix with the 
same number of columns as \code{y}, depending on the model.}

\item{bw}{Integer \code{(>= 0)}. The bandwidth parameter controlling smoothing 
of the moment conditions. Defaults to "auto" in which case the optimal parameter 
value is chosen according to \code{bwAndrews()}. For highly autocorrelated moment 
conditions higher values of \code{bw} are needed, while for independent data this 
should be set to zero, in which case no smoothing takes place. For more, see details.}

\item{initial}{A function that returns an initial estimate for the parameters 
and their covariance matrix as a list with two elements called \code{th} and \code{cov}. 
By default OLS estimates are used and should be used whenever possible.}

\item{args}{A list containing optional further arguments to \code{g()} and 
\code{prior_fun()}. For a var model, the lag length can be set for example 
by \code{args = list(p = 5)}. For more, see details.}

\item{verbose}{Logical. Defaults to \code{TRUE} in which case messages are produced.}
}
\value{
\code{init_sbetel()} returns a list that defines a sbetel model 
and can be passed on to \code{eval_sbetel()} or \code{est_sbetel()}.
}
\description{
Initialize a smoothed BETEL model by providing the model defining 
function \code{g()}, the prior generating function \code{prior_fun()} and the 
data \code{y} to be used for the estimation of the model. Alternatively 
ready-made functions can be used (e.g. by setting \code{g = "var"} and 
\code{prior_fun() = "var"}). \code{init_sbetel()} returns 
a list that defines the model and can be passed on either to \code{eval_sbetel()} or 
\code{est_sbetel()}.

#' @param y A data matrix. Input for \code{g()} below. E.g. for \code{g = "var"} this
would be a \code{T x k} matrix with \code{T} observations of \code{k} variables.
}
\details{
TBA
}
\examples{
\dontrun{
model <- init_sbetel(y = y, bw = 1, args = list(p = 2))
}
}
