\name{sbetel-package}
\alias{sbetel-package}
\alias{sbetel}
\docType{package}
\title{
\packageTitle{sbetel}
}
\description{
\packageDescription{sbetel}
}
\details{
...
}
\author{
\packageAuthor{sbetel}

Maintainer: \packageMaintainer{sbetel}
}
\references{
...
}
\keyword{ package }
\seealso{
  ...
}
\examples{
  ...
}
