% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{eval_sbetel}
\alias{eval_sbetel}
\title{Evaluate a smoothed BETEL posterior density}
\usage{
eval_sbetel(th, model, itermax = 20)
}
\arguments{
\item{th}{A numerical vector of parameter values.}

\item{model}{A list returned by \code{init_sbetel()}, defining the model.}

\item{itermax}{Maximum number of Newton-Rhapson iterations within the evaluation of the likelihood.}
}
\value{
\code{eval_sbetel()} returns a numerical value that is the unnormalized
posterior density at \code{th}.
}
\description{
Evaluate a (unnormalized) smoothed BETEL posterior density at parameter
values \code{th} given \code{model} (list) constructed by \code{init_sbetel()}.
}
\examples{
eval_sbetel(model$th_initial, model)
}
