% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{eval_sbetel}
\alias{eval_sbetel}
\title{Evaluate a smoothed BETEL density}
\usage{
eval_sbetel(th, model, xy = model$args$xy, td = 0, itermax = 20)
}
\arguments{
\item{th}{A numerical vector of parameter values.}

\item{model}{A list returned by \code{init_sbetel()} defining the model.}

\item{xy}{The data for which the density is coniditoned on. Defaults to the 
data passed to \code{init_sbetel()} without the prior.}

\item{td}{The number of rows of prior data, i.e. the number of rows from 
the beginning of the data that should not be smoothed.}

\item{itermax}{Maximum number of Newton-Rhapson iterations within the evaluation 
of the likelihood. Defaults to 20 which should be more than enough.}
}
\value{
\code{eval_sbetel()} returns a numerical value that is the unnormalized 
density at \code{th} conditional on \code{xy}.
}
\description{
Evaluate a (unnormalized) smoothed BETEL density at parameter
values \code{th} given the \code{model} (list) generated by \code{init_sbetel()} 
and the data \code{xy}. 
A wrapper for c++ implementation.
}
\examples{
\dontrun{
eval_sbetel(th = model$initial(model$args$xy)$th, model = model, xy = model$args$xy)
}
}
