% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{init_sbetel}
\alias{init_sbetel}
\title{Initialize a smoothed BETEL model}
\usage{
init_sbetel(
  y,
  p = "auto",
  lambda = "auto",
  bw = "auto",
  stat = rep(0, ncol(y)),
  sigma = FALSE,
  additional_priors = NULL
)
}
\arguments{
\item{y}{\code{(T x n)} numerical matrix with \code{n} time series
in its columns.}

\item{p}{Lag length. Defaults to \code{"auto"} in which case the optimal
lag length is chosen according to the out-of-sample mean squared 
forecasting error of the joint posterior mode.}

\item{lambda}{The inverse amount of shrinkage towards the Minnesotaprior \code{(> 0)}. 
Defaults to \code{"auto"} in which case the optimal shrinkage is chosen according to the 
out-of-sample mean squared forecasting error of the joint posterior mode. 
No shrinkage is obtained by setting this to \code{Inf}.}

\item{bw}{Integer \code{(>= 0)}. The bandwidth parameter used for smoothing of the moment 
conditions. Defaults to \code{"auto"} in which case the optimal value is 
chosen according to \code{bwAndrews()} from \code{sandwich} package.}

\item{stat}{A numerical vector of length \code{n} \code{(0 <= n < 1)}. 
The prior mean of the coefficients for own first lags in the VAR model. 
For series with a priori more persistent dynamics, values closer to zero 
can be chosen. Defaults to zero for all series.}

\item{sigma}{Logical. Defaults to \code{FALSE}. If set to \code{TRUE}, the 
Cholesky decomposition of the covariance of the errors is estimated.}

\item{additional_priors}{\code{NULL} or numerical vector of length 2, where
both elements \code{> 0}. First element controls the inverse weight of
\emph{sum-of-coefficients} (SOC) prior whereas the other controls the 
\emph{initial-dummy-observation} (IDO) prior. Typical values in the literature 
for these values would be \code{c(1,1)}. Defaults to \code{NULL} in which case 
these priors are not used.}
}
\value{
\code{init_sbetel()} returns a list that defines the model and can be passed
to \code{eval_sbetel()} or \code{est_sbetel}.
}
\description{
Initialize a smoothed (VAR) BETEL model by providing the data and 
the desired parameters. This function returns a list that defines
the model and can be passed either to \code{eval_sbetel()} or 
\code{est_sbetel()}.
}
\examples{
model <- init_sbetel(y)
}
