% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{init_sbetel}
\alias{init_sbetel}
\title{Initialize a smoothed BETEL model}
\usage{
init_sbetel(
  g = "var",
  y,
  bw = "auto",
  args = list(sigma = TRUE, gmm = FALSE),
  initial = list(th = NULL, cov = NULL),
  verbose = TRUE
)
}
\arguments{
\item{g}{A function of the form \code{g(th, y, args)} defining the model.
Inputs should include a numerical parameter vector \code{th}, a data matrix 
\code{y} and optional further arguments through a list called \code{args}.
Output shoud be a numerical matrix with its columns corresponding to the sample
moment conditions of the model. Defaults to \code{"var"} in which case a 
ready-made function for vector autoregressive models is used. For more, 
see details.}

\item{y}{A data matrix. Input for \code{g()} above. E.g. for \code{g = "var"} this
would be a \code{T x k} matrix with \code{T} observations of \code{k} variables.}

\item{bw}{Integer \code{(>= 0)}. The bandwidth parameter controlling smoothing of the moment 
conditions. Defaults to "auto" in which case the optimal parameter value is 
chosen according to \code{bwAndrews()}. For more, see details.}

\item{args}{A list containing optional further arguments to \code{g()}.
If \code{g = "var"} is chosen, this can be left undefined in which case the 
parameters of the var model are automatically chosen to minimize the 
out-of-sample squared forecasting errors of the model. For example, the lag 
length and the shrinkage parameter can be manually fixed by defining 
\code{args = list(p = 5, lambda = 0.2)}. For more, see details.}

\item{initial}{A list of initial parameter values that defaults to 
\code{list(th = NULL, cov = NULL)}. At least \code{initial$th} must be predefined, 
if \code{g} is not set as \code{"var"}. If \code{initial$cov} is not predefined, 
it is approximated with a two-step GMM.}

\item{verbose}{Logical. Defaults to \code{TRUE} in which case messages are produced.}
}
\value{
\code{init_sbetel()} returns a list that defines a sbetel model 
and can be passed on to \code{eval_sbetel()} or \code{est_sbetel()}.
}
\description{
Initialize a smoothed BETEL model by providing the model defining 
function \code{g()} and the data \code{y} to be used for estimation of the model. Alternatively 
ready-made functions can be used (e.g. by setting \code{g = "var"}). \code{init_sbetel()} returns 
a list that defines the model and that can be passed on either to \code{eval_sbetel()} or 
\code{est_sbetel()}.
}
\details{
TBA
}
\examples{
set.seed(222)
y <- matrix(rnorm(120), ncol = 3)
model <- init_sbetel(g = "var", y = y)
}
