\name{sbetel-package}
\alias{sbetel-package}
\alias{sbetel}
\docType{package}
\title{
\packageTitle{sbetel}
}
\description{
\packageDescription{sbetel}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{sbetel}
\packageIndices{sbetel}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{sbetel}

Maintainer: \packageMaintainer{sbetel}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
